Improve transitions and grass type distribution

+++ Assets/Scripts/GrassSimulation/Core/Inputs/GrassMapInput.cs
@@ -8,7 +8,7 @@ namespace GrassSimulation.Core.Inputs
 		///     <para>Gets the type of grass at given position.</para>
 		/// </summary>
 		/// <returns>The index corresponding to the array position in the GrassContainer represtenting the grass type.</returns> 
-		public abstract byte GetGrassType(float x, float y, float z);
+		public abstract int GetGrassType(float x, float y, float z);
 
 		
 		/// <summary>
